ecommerce-ai-agent/requirements.txt
# Delete the old one and create a new one

python-dotenv==1.0.0
pandas>=2.2.0
numpy>=1.26.0
langchain==0.2.0
langgraph==0.2.0
langchain-community==0.2.0
langchain-huggingface==0.0.3
huggingface_hub>=0.22.0
transformers>=4.39.0
sentence-transformers>=2.6.0
prophet==1.1.5
scikit-learn>=1.3.0
joblib>=1.3.0
mlflow==2.10.0
sqlalchemy>=2.0.0
streamlit>=1.32.0
plotly>=5.18.0
requests>=2.31.0
loguru==0.7.0